+++ src/main/java/io/github/tr100000/descriptor/gui/MobEffectTooltipCache.java
@@ -7,9 +7,18 @@ import net.minecraft.world.effect.MobEffectInstance;
 import java.util.Map;
 
 public class MobEffectTooltipCache {
+    private final MobEffectTooltipSettings settings;
     private final Map<MobEffectInstance, ClientTooltipComponent> cache = new Object2ObjectOpenHashMap<>();
 
-    public ClientTooltipComponent getOrCreate(MobEffectInstance effectInstance) {
-        return cache.computeIfAbsent(effectInstance, MobEffectTooltipComponent::new);
+    public MobEffectTooltipCache() {
+        this(MobEffectTooltipSettings.DEFAULT);
+    }
+
+    public MobEffectTooltipCache(MobEffectTooltipSettings settings) {
+        this.settings = settings;
+    }
+
+    public ClientTooltipComponent getOrCreate(MobEffectInstance instance) {
+        return cache.computeIfAbsent(instance, i -> new MobEffectTooltipComponent(i, settings));
     }
 }