+++ src/main/java/io/github/tr100000/descriptor/gui/MobEffectTooltipComponent.java
@@ -7,29 +7,55 @@ import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.network.chat.CommonComponents;
 import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.Identifier;
 import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffectUtil;
 
 public class MobEffectTooltipComponent implements ClientTooltipComponent {
     private static final Minecraft minecraft = Minecraft.getInstance();
-    private final Component nameComponent;
+
+    private final MobEffectInstance instance;
+    private final MobEffectTooltipSettings settings;
+
+    private final MutableComponent nameComponent;
     private final Component descriptionComponent;
     private final Component idComponent;
 
-    public MobEffectTooltipComponent(MobEffectInstance effectInstance) {
-        Identifier id = BuiltInRegistries.MOB_EFFECT.getKey(effectInstance.getEffect().value());
+    private Component lastShownNameComponent;
+
+    public MobEffectTooltipComponent(MobEffectInstance instance, MobEffectTooltipSettings settings) {
+        this.instance = instance;
+        this.settings = settings;
+
+        Identifier id = BuiltInRegistries.MOB_EFFECT.getKey(instance.getEffect().value());
         assert id != null;
 
         nameComponent = Component.translatable(id.toLanguageKey("effect"));
-        descriptionComponent = DescriptorUtil.getMobEffectDescription(effectInstance.getEffect()).withStyle(ChatFormatting.GRAY);
+        descriptionComponent = DescriptorUtil.getMobEffectDescription(instance.getEffect()).withStyle(ChatFormatting.GRAY);
         idComponent = Component.literal(id.toString()).withStyle(ChatFormatting.DARK_GRAY);
+
+        if (settings.showAmplifier() && instance.getAmplifier() >= 1 && instance.getAmplifier() <= 9) {
+            nameComponent.append(CommonComponents.SPACE).append(Component.translatable("enchantment.level." + (instance.getAmplifier() + 1)));
+        }
+
+        lastShownNameComponent = settings.showDuration() ? getNameComponentWithDuration() : nameComponent;
     }
 
     @Override
     public void extractText(GuiGraphicsExtractor graphics, Font font, int x, int y) {
-        graphics.text(font, nameComponent, x, y, -1);
+        if (settings.showDuration()) {
+            lastShownNameComponent = getNameComponentWithDuration();
+            graphics.text(font, lastShownNameComponent, x, y, -1);
+        }
+        else {
+            graphics.text(font, nameComponent, x, y, -1);
+        }
+
         graphics.text(font, descriptionComponent, x, y + 10, -1);
+
         if (minecraft.options.advancedItemTooltips) {
             graphics.text(font, idComponent, x, y + 20, -1);
         }
@@ -42,11 +68,19 @@ public class MobEffectTooltipComponent implements ClientTooltipComponent {
 
     @Override
     public int getWidth(Font font) {
-        int longest = font.width(nameComponent);
+        int longest = font.width(lastShownNameComponent);
         longest = Math.max(longest, font.width(descriptionComponent));
         if (minecraft.options.advancedItemTooltips) {
             longest = Math.max(longest, font.width(idComponent));
         }
         return longest;
+    }
+
+    private Component getNameComponentWithDuration() {
+        assert minecraft.level != null;
+        return nameComponent.copy()
+                .append(" (")
+                .append(MobEffectUtil.formatDuration(instance, 1.0F, minecraft.level.tickRateManager().tickrate()))
+                .append(")");
     }
 }