+++ src/main/java/io/github/tr100000/descriptor/gui/MobEffectTooltipSettings.java
@@ -1,0 +1,8 @@
+package io.github.tr100000.descriptor.gui;
+
+public record MobEffectTooltipSettings(
+        boolean showAmplifier,
+        boolean showDuration
+) {
+    public static final MobEffectTooltipSettings DEFAULT = new MobEffectTooltipSettings(false, false);
+}