+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### Added
 - The mod is now configurable!
   - Install ModMenu and YACL to edit the config in-game
+- Now shows effect amplifier and remaining duration in HUD tooltips
+- Added an option to show effect amplifier and remaining duration in tooltips show in the inventory effects